requirements.txt
Flask==3.1.0
ipython==8.31.0
python-decouple==3.8